+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## Release 0.3
+
+Handle modules with ':latest' version.
+
 ## Release 0.2
 
-Added configuration file and proxy support
+Added configuration file and proxy support.
 
 ## Release 0.1
 
-Initial version
+Initial version.